cpu-o3: split decoupled bpu stats helpers
moving DB/init/dumpStats etc to new file to reduce clutter.
