+++ src/cpu/pred/SConscript
@@ -95,6 +95,7 @@ Source('ftb/folded_hist.cc')
 Source('ftb/ras.cc')
 Source('ftb/uras.cc')
 Source('btb/decoupled_bpred.cc')
+Source('btb/decoupled_bpred_stats.cc')
 Source('btb/abtb.cc')
 Source('btb/mbtb.cc')
 Source('btb/timed_base_pred.cc')
